requirements.txt
matplotlib>=3.8.2
numpy>=1.26.1
Pillow>=10.1.0
seaborn>=0.13.2
tqdm>=4.66.2